reductor-module/reductor-service-v2/requirements.txt
"""
requirements.txt

Reductor Service v2 dependencies
"""

fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
lxml==4.9.3
pdf2docx==0.5.1
minio==7.2.0
pymupdf==1.23.8
python-dotenv==1.0.0
